+++ src/fidesops/ops/schemas/privacy_request.py
@@ -9,9 +9,9 @@ from pydantic import Field, validator
 from fidesops.ops.core.config import config
 from fidesops.ops.models.policy import ActionType
 from fidesops.ops.models.privacy_request import (
+    CollectionActionRequired,
     ExecutionLogStatus,
     PrivacyRequestStatus,
-    StoppedCollection,
 )
 from fidesops.ops.schemas.api import BulkResponse, BulkUpdateFailed
 from fidesops.ops.schemas.base_class import BaseSchema
@@ -133,7 +133,7 @@ class RowCountRequest(BaseSchema):
     row_count: int
 
 
-class StoppedCollectionDetails(StoppedCollection):
+class CollectionActionRequiredDetails(CollectionActionRequired):
 
     collection: Optional[str] = None  # type: ignore
 
@@ -164,7 +164,7 @@ class PrivacyRequestResponse(BaseSchema):
     # creation.
     identity: Optional[Dict[str, Optional[str]]]
     policy: PolicySchema
-    stopped_collection_details: Optional[StoppedCollectionDetails] = None
+    stopped_collection_details: Optional[CollectionActionRequiredDetails] = None
     resume_endpoint: Optional[str]
 
     class Config: