+++ src/fidesops/ops/task/graph_task.py
@@ -492,7 +492,7 @@ class GraphTask(ABC):  # pylint: disable=too-many-instance-attributes
         return filtered_output
 
     @retry(action_type=ActionType.erasure, default_return=0)
-    def erasure_request(self, retrieved_data: List[Row]) -> int:
+    def erasure_request(self, retrieved_data: List[Row], *inputs: List[Row]) -> int:
         """Run erasure request"""
         # if there is no primary key specified in the graph node configuration
         # note this in the execution log and perform no erasures on this node
@@ -523,11 +523,16 @@ class GraphTask(ABC):  # pylint: disable=too-many-instance-attributes
             )
             return 0
 
+        formatted_input_data: NodeInput = self.pre_process_input_data(
+            *inputs, group_dependent_fields=True
+        )
+
         output = self.connector.mask_data(
             self.traversal_node,
             self.resources.policy,
             self.resources.request,
             retrieved_data,
+            formatted_input_data,
         )
         self.log_end(ActionType.erasure)
         self.resources.cache_erasure(
@@ -701,8 +706,23 @@ async def run_erasure(  # pylint: disable = too-many-arguments, too-many-locals
             The termination function just returns this tuple of ints."""
             return dependent_values
 
+        access_request_data[ROOT_COLLECTION_ADDRESS.value] = [identity]
+
         dsk: Dict[CollectionAddress, Any] = {
-            k: (t.erasure_request, access_request_data[str(k)]) for k, t in env.items()
+            k: (
+                t.erasure_request,
+                access_request_data[
+                    str(k)
+                ],  # Pass in the results of the access request for this collection
+                *[
+                    access_request_data[
+                        str(upstream_key)
+                    ]  # Additionally pass in the original input data we used for the access request. It's helpful in
+                    # cases like the EmailConnector where the access request doesn't actually retrieve data.
+                    for upstream_key in t.input_keys
+                ],
+            )
+            for k, t in env.items()
         }
         # terminator function waits for all keys
         dsk[TERMINATOR_ADDRESS] = (termination_fn, *env.keys())