+++ tests/ops/api/v1/endpoints/test_privacy_request_endpoints.py
@@ -49,7 +49,7 @@ from fidesops.ops.email_templates import get_email_template
 from fidesops.ops.graph.config import CollectionAddress
 from fidesops.ops.graph.graph import DatasetGraph
 from fidesops.ops.models.datasetconfig import DatasetConfig
-from fidesops.ops.models.policy import ActionType, PausedStep
+from fidesops.ops.models.policy import ActionType, CurrentStep
 from fidesops.ops.models.privacy_request import (
     ExecutionLog,
     ExecutionLogStatus,
@@ -60,12 +60,10 @@ from fidesops.ops.models.privacy_request import (
 from fidesops.ops.schemas.dataset import DryRunDatasetResponse
 from fidesops.ops.schemas.email.email import (
     EmailActionType,
-    EmailForActionType,
     SubjectIdentityVerificationBodyParams,
 )
 from fidesops.ops.schemas.masking.masking_secrets import SecretType
 from fidesops.ops.schemas.policy import PolicyResponse
-from fidesops.ops.schemas.privacy_request import ExecutionAndAuditLogResponse
 from fidesops.ops.schemas.redis_cache import PrivacyRequestIdentity
 from fidesops.ops.util.cache import (
     get_encryption_cache_key,
@@ -1180,7 +1178,7 @@ class TestGetPrivacyRequests:
         # Mock the privacy request being in a paused state waiting for manual input to the "manual_collection"
         privacy_request.status = PrivacyRequestStatus.paused
         privacy_request.save(db)
-        paused_step = PausedStep.access
+        paused_step = CurrentStep.access
         paused_collection = CollectionAddress("manual_dataset", "manual_collection")
         privacy_request.cache_paused_collection_details(
             step=paused_step,
@@ -1221,7 +1219,7 @@ class TestGetPrivacyRequests:
         # Mock the privacy request being in a paused state waiting for manual erasure confirmation to the "another_collection"
         privacy_request.status = PrivacyRequestStatus.paused
         privacy_request.save(db)
-        paused_step = PausedStep.erasure
+        paused_step = CurrentStep.erasure
         paused_collection = CollectionAddress("manual_dataset", "another_collection")
         privacy_request.cache_paused_collection_details(
             step=paused_step,
@@ -1280,7 +1278,7 @@ class TestGetPrivacyRequests:
         privacy_request.status = PrivacyRequestStatus.error
         privacy_request.save(db)
         privacy_request.cache_failed_collection_details(
-            step=PausedStep.erasure,
+            step=CurrentStep.erasure,
             collection=CollectionAddress("manual_example", "another_collection"),
         )
 
@@ -2087,7 +2085,7 @@ class TestResumeAccessRequestWithManualInput:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.access,
+            step=CurrentStep.access,
             collection=CollectionAddress("manual_example", "filing_cabinet"),
         )
 
@@ -2117,7 +2115,7 @@ class TestResumeAccessRequestWithManualInput:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.access,
+            step=CurrentStep.access,
             collection=CollectionAddress("manual_input", "filing_cabinet"),
         )
 
@@ -2150,7 +2148,7 @@ class TestResumeAccessRequestWithManualInput:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.access,
+            step=CurrentStep.access,
             collection=CollectionAddress("manual_input", "filing_cabinet"),
         )
 
@@ -2273,7 +2271,7 @@ class TestResumeErasureRequestWithManualConfirmation:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.erasure,
+            step=CurrentStep.erasure,
             collection=CollectionAddress("manual_example", "filing_cabinet"),
         )
 
@@ -2295,7 +2293,7 @@ class TestResumeErasureRequestWithManualConfirmation:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.access,
+            step=CurrentStep.access,
             collection=CollectionAddress("manual_example", "filing_cabinet"),
         )
         response = api_client.post(url, headers=auth_header, json={"row_count": 0})
@@ -2328,7 +2326,7 @@ class TestResumeErasureRequestWithManualConfirmation:
         privacy_request.save(db)
 
         privacy_request.cache_paused_collection_details(
-            step=PausedStep.erasure,
+            step=CurrentStep.erasure,
             collection=CollectionAddress("manual_input", "filing_cabinet"),
         )
         response = api_client.post(
@@ -2402,7 +2400,7 @@ class TestRestartFromFailure:
         privacy_request.save(db)
 
         privacy_request.cache_failed_collection_details(
-            step=PausedStep.access,
+            step=CurrentStep.access,
             collection=CollectionAddress("test_dataset", "test_collection"),
         )
 
@@ -2414,7 +2412,7 @@ class TestRestartFromFailure:
 
         submit_mock.assert_called_with(
             privacy_request_id=privacy_request.id,
-            from_step=PausedStep.access.value,
+            from_step=CurrentStep.access.value,
             from_webhook_id=None,
         )