+++ CHANGELOG.md
@@ -23,6 +23,13 @@ The types of changes are:
 
 * Created a docker image for the privacy center [1165](https://github.com/ethyca/fidesops/pull/1165)
 
+
+### Added
+
+* Adds users and owners configuration for Hubspot connector [#1091](https://github.com/ethyca/fidesops/pull/1091)
+* Foundations for a new email connector type [#1142](https://github.com/ethyca/fidesops/pull/1142)
+* Have the new email connector cache action needed for each collection [#1168](https://github.com/ethyca/fidesops/pull/1168)
+
 ### Docs
 
 * Fix analytics opt out environment variable name [#1170](https://github.com/ethyca/fidesops/pull/1170)
@@ -33,7 +40,6 @@ The types of changes are:
 
 * Fix issue with fideslog event loop errors [#1174](https://github.com/ethyca/fidesops/pull/1174)
 * Allow passwords to be sent either base64 encode or as plaintext. [#1236](https://github.com/ethyca/fidesops/pull/1236)
-
 
 ## [1.7.2](https://github.com/ethyca/fidesops/compare/1.7.1...1.7.2)